Expense Bot/requirements.txt
python-telegram-bot==13.15
gspread==5.12.0
google-auth==2.35.0
